+++ app/src/androidTest/kotlin/com/mgaetan89/showsrage/extension/realm/RealmExtension_ClearScheduleTest.kt
@@ -1,29 +1,15 @@
 package com.mgaetan89.showsrage.extension.realm
 
-import android.os.Looper
-import android.support.test.rule.ActivityTestRule
 import android.support.test.runner.AndroidJUnit4
-import com.mgaetan89.showsrage.TestActivity
 import com.mgaetan89.showsrage.extension.clearSchedule
 import com.mgaetan89.showsrage.model.Schedule
-import io.realm.Realm
 import org.assertj.core.api.Assertions.assertThat
-import org.junit.After
-import org.junit.AfterClass
 import org.junit.Before
-import org.junit.BeforeClass
-import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
-class RealmExtension_ClearScheduleTest {
-    @JvmField
-    @Rule
-    val activityRule = ActivityTestRule(TestActivity::class.java)
-
-    private val realm: Realm by lazy { Realm.getDefaultInstance() }
-
+class RealmExtension_ClearScheduleTest : RealmTest() {
     @Before
     fun before() {
         this.realm.isAutoRefresh = false
@@ -38,26 +24,5 @@ class RealmExtension_ClearScheduleTest {
         assertThat(this.getSchedule()).isEmpty()
     }
 
-    @After
-    fun after() {
-        this.realm.close()
-    }
-
     private fun getSchedule() = this.realm.where(Schedule::class.java).findAll()
-
-    companion object {
-        @BeforeClass
-        @JvmStatic
-        fun beforeClass() {
-            if (Looper.myLooper() == null) {
-                Looper.prepare()
-            }
-        }
-
-        @AfterClass
-        @JvmStatic
-        fun afterClass() {
-            Looper.myLooper().quit()
-        }
-    }
 }
